# openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "blueeconomy-insurance",
  "version": "0.1.0"
 },
 "paths": {
...
  "/v1/claims": {
   "post": {
    "tags": [
     "claims"
    ],
    "summary": "File Fnol",
    "operationId": "file_fnol_v1_claims_post",
    "parameters": [
     {
      "name": "Idempotency-Key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Idempotency-Key"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/FnolIn"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response File Fnol V1 Claims Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/claims/{claim_ref}": {
   "get": {
    "tags": [
     "claims"
    ],
    "summary": "Get Claim",
    "operationId": "get_claim_v1_claims__claim_ref__get",
    "parameters": [
     {
      "name": "claim_ref",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Claim Ref"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Get Claim V1 Claims  Claim Ref  Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/claims/{claim_ref}/documents": {
   "post": {
    "tags": [
     "claims"
    ],
    "summary": "Attach Document",
    "operationId": "attach_document_v1_claims__claim_ref__documents_post",
    "parameters": [
     {
      "name": "claim_ref",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Claim Ref"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/DocumentIn"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Attach Document V1 Claims  Claim Ref  Documents Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/claims/{claim_ref}:propose-adjuster": {
   "post": {
    "tags": [
     "claims"
    ],
    "summary": "Propose Adjuster",
    "operationId": "propose_adjuster_v1_claims__claim_ref__propose_adjuster_post",
    "parameters": [
     {
      "name": "claim_ref",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Claim Ref"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AdjusterProposal"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Propose Adjuster V1 Claims  Claim Ref  Propose Adjuster Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/claims/{claim_ref}:confirm-adjuster": {
   "post": {
    "tags": [
     "claims"
    ],
    "summary": "Confirm Adjuster",
    "operationId": "confirm_adjuster_v1_claims__claim_ref__confirm_adjuster_post",
    "parameters": [
     {
      "name": "claim_ref",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Claim Ref"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Confirm Adjuster V1 Claims  Claim Ref  Confirm Adjuster Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/claims/{claim_ref}:propose-settlement": {
   "post": {
    "tags": [
     "claims"
    ],
    "summary": "Propose Settlement",
    "operationId": "propose_settlement_v1_claims__claim_ref__propose_settlement_post",
    "parameters": [
     {
      "name": "claim_ref",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Claim Ref"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SettlementProposal"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Propose Settlement V1 Claims  Claim Ref  Propose Settlement Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/claims/{claim_ref}:approve-settlement": {
   "post": {
    "tags": [
     "claims"
    ],
    "summary": "Approve Settlement",
    "operationId": "approve_settlement_v1_claims__claim_ref__approve_settlement_post",
    "parameters": [
     {
      "name": "claim_ref",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Claim Ref"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Approve Settlement V1 Claims  Claim Ref  Approve Settlement Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/claims/{claim_ref}:reject": {
   "post": {
    "tags": [
     "claims"
    ],
    "summary": "Reject Claim",
    "operationId": "reject_claim_v1_claims__claim_ref__reject_post",
    "parameters": [
     {
      "name": "claim_ref",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Claim Ref"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RejectIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Reject Claim V1 Claims  Claim Ref  Reject Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
   "AdjusterProposal": {
    "properties": {
     "adjuster_sub": {
      "type": "string",
      "maxLength": 256,
      "minLength": 1,
      "title": "Adjuster Sub"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "adjuster_sub"
    ],
    "title": "AdjusterProposal"
   },
...
   "DocumentIn": {
    "properties": {
     "vault_ref": {
      "type": "string",
      "maxLength": 256,
      "minLength": 1,
      "title": "Vault Ref"
     },
     "sha256": {
      "type": "string",
      "pattern": "^[0-9a-f]{64}$",
      "title": "Sha256"
     },
     "description": {
      "type": "string",
      "maxLength": 256,
      "title": "Description",
      "default": ""
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "vault_ref",
     "sha256"
    ],
    "title": "DocumentIn"
   },
...
   "FnolIn": {
    "properties": {
     "policy_number": {
      "type": "string",
      "maxLength": 40,
      "minLength": 1,
      "title": "Policy Number"
     },
     "loss_occurred_at": {
      "type": "string",
      "format": "date-time",
      "title": "Loss Occurred At"
     },
     "loss_description": {
      "type": "string",
      "maxLength": 4000,
      "title": "Loss Description",
      "default": ""
     },
     "claimed_kobo": {
      "type": "integer",
      "exclusiveMinimum": 0.0,
      "title": "Claimed Kobo"
     },
     "trigger_event_id": {
      "type": "string",
      "maxLength": 128,
      "title": "Trigger Event Id",
      "default": ""
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "policy_number",
     "loss_occurred_at",
     "claimed_kobo"
    ],
    "title": "FnolIn"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "RejectIn": {
    "properties": {
     "reason": {
      "type": "string",
      "maxLength": 2000,
      "minLength": 1,
      "title": "Reason"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "reason"
    ],
    "title": "RejectIn"
   },
   "SettlementProposal": {
    "properties": {
     "settled_kobo": {
      "type": "integer",
      "exclusiveMinimum": 0.0,
      "title": "Settled Kobo"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "settled_kobo"
    ],
    "title": "SettlementProposal"
   },
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   }
  }
 }
}